Minor compiling fixes

+++ tests/MqttClientTest.cpp
@@ -199,8 +199,8 @@ static int s_TestMqtt311DirectConnectionMinimal(Aws::Crt::Allocator *allocator, 
     Aws::Crt::ApiHandle apiHandle(allocator);
     Aws::Crt::Mqtt::MqttClient client;
     Aws::Crt::Io::SocketOptions socketOptions;
-    std::shared_ptr<Aws::Crt::Mqtt::MqttConnection> connection =
-        client.NewConnection(aws_string_c_str(endpoint), std::stoi(aws_string_c_str(port)), socketOptions, false);
+    std::shared_ptr<Aws::Crt::Mqtt::MqttConnection> connection = client.NewConnection(
+        aws_string_c_str(endpoint), (uint16_t)std::stoi(aws_string_c_str(port)), socketOptions, false);
     int connectResult = s_ConnectAndDisconnect(connection);
     if (connectResult != AWS_OP_SUCCESS)
     {
@@ -248,8 +248,8 @@ static int s_TestMqtt311DirectConnectionWithBasicAuth(Aws::Crt::Allocator *alloc
     Aws::Crt::ApiHandle apiHandle(allocator);
     Aws::Crt::Mqtt::MqttClient client;
     Aws::Crt::Io::SocketOptions socketOptions;
-    std::shared_ptr<Aws::Crt::Mqtt::MqttConnection> connection =
-        client.NewConnection(aws_string_c_str(endpoint), std::stoi(aws_string_c_str(port)), socketOptions, false);
+    std::shared_ptr<Aws::Crt::Mqtt::MqttConnection> connection = client.NewConnection(
+        aws_string_c_str(endpoint), (uint16_t)std::stoi(aws_string_c_str(port)), socketOptions, false);
     connection->SetLogin(aws_string_c_str(username), aws_string_c_str(password));
     int connectResult = s_ConnectAndDisconnect(connection);
     if (connectResult != AWS_OP_SUCCESS)
@@ -299,7 +299,7 @@ static int s_TestMqtt311DirectConnectionWithTLS(Aws::Crt::Allocator *allocator, 
     Aws::Crt::Mqtt::MqttClient client;
     Aws::Crt::Io::SocketOptions socketOptions;
     std::shared_ptr<Aws::Crt::Mqtt::MqttConnection> connection = client.NewConnection(
-        aws_string_c_str(endpoint), std::stoi(aws_string_c_str(port)), socketOptions, tlsContext, false);
+        aws_string_c_str(endpoint), (uint16_t)std::stoi(aws_string_c_str(port)), socketOptions, tlsContext, false);
     int connectResult = s_ConnectAndDisconnect(connection);
     if (connectResult != AWS_OP_SUCCESS)
     {
@@ -406,13 +406,13 @@ static int s_TestMqtt311DirectConnectionWithHttpProxy(Aws::Crt::Allocator *alloc
 
     Aws::Crt::Http::HttpClientConnectionProxyOptions proxyOptions;
     proxyOptions.HostName = aws_string_c_str(proxy_endpoint);
-    proxyOptions.Port = std::stoi(aws_string_c_str(proxy_port));
+    proxyOptions.Port = (uint16_t)std::stoi(aws_string_c_str(proxy_port));
     proxyOptions.ProxyConnectionType = Aws::Crt::Http::AwsHttpProxyConnectionType::Tunneling;
 
     Aws::Crt::Mqtt::MqttClient client;
     Aws::Crt::Io::SocketOptions socketOptions;
     std::shared_ptr<Aws::Crt::Mqtt::MqttConnection> connection = client.NewConnection(
-        aws_string_c_str(endpoint), std::stoi(aws_string_c_str(port)), socketOptions, tlsContext, false);
+        aws_string_c_str(endpoint), (uint16_t)std::stoi(aws_string_c_str(port)), socketOptions, tlsContext, false);
     connection->SetHttpProxyOptions(proxyOptions);
     int connectResult = s_ConnectAndDisconnect(connection);
     if (connectResult != AWS_OP_SUCCESS)
@@ -459,8 +459,8 @@ static int s_TestMqtt311WSConnectionMinimal(Aws::Crt::Allocator *allocator, void
     Aws::Crt::ApiHandle apiHandle(allocator);
     Aws::Crt::Mqtt::MqttClient client;
     Aws::Crt::Io::SocketOptions socketOptions;
-    std::shared_ptr<Aws::Crt::Mqtt::MqttConnection> connection =
-        client.NewConnection(aws_string_c_str(endpoint), std::stoi(aws_string_c_str(port)), socketOptions, true);
+    std::shared_ptr<Aws::Crt::Mqtt::MqttConnection> connection = client.NewConnection(
+        aws_string_c_str(endpoint), (uint16_t)std::stoi(aws_string_c_str(port)), socketOptions, true);
     int connectResult = s_ConnectAndDisconnect(connection);
     if (connectResult != AWS_OP_SUCCESS)
     {
@@ -507,8 +507,8 @@ static int s_TestMqtt311WSConnectionWithBasicAuth(Aws::Crt::Allocator *allocator
     Aws::Crt::ApiHandle apiHandle(allocator);
     Aws::Crt::Mqtt::MqttClient client;
     Aws::Crt::Io::SocketOptions socketOptions;
-    std::shared_ptr<Aws::Crt::Mqtt::MqttConnection> connection =
-        client.NewConnection(aws_string_c_str(endpoint), std::stoi(aws_string_c_str(port)), socketOptions, true);
+    std::shared_ptr<Aws::Crt::Mqtt::MqttConnection> connection = client.NewConnection(
+        aws_string_c_str(endpoint), (uint16_t)std::stoi(aws_string_c_str(port)), socketOptions, true);
     connection->SetLogin(aws_string_c_str(username), aws_string_c_str(password));
     int connectResult = s_ConnectAndDisconnect(connection);
     if (connectResult != AWS_OP_SUCCESS)
@@ -557,7 +557,7 @@ static int s_TestMqtt311WSConnectionWithTLS(Aws::Crt::Allocator *allocator, void
     Aws::Crt::Mqtt::MqttClient client;
     Aws::Crt::Io::SocketOptions socketOptions;
     std::shared_ptr<Aws::Crt::Mqtt::MqttConnection> connection = client.NewConnection(
-        aws_string_c_str(endpoint), std::stoi(aws_string_c_str(port)), socketOptions, tlsContext, true);
+        aws_string_c_str(endpoint), (uint16_t)std::stoi(aws_string_c_str(port)), socketOptions, tlsContext, true);
     int connectResult = s_ConnectAndDisconnect(connection);
     if (connectResult != AWS_OP_SUCCESS)
     {
@@ -608,14 +608,14 @@ static int s_TestMqtt311WSConnectionWithHttpProxy(Aws::Crt::Allocator *allocator
 
     Aws::Crt::Http::HttpClientConnectionProxyOptions proxyOptions;
     proxyOptions.HostName = aws_string_c_str(proxy_endpoint);
-    proxyOptions.Port = std::stoi(aws_string_c_str(proxy_port));
+    proxyOptions.Port = (uint16_t)std::stoi(aws_string_c_str(proxy_port));
     proxyOptions.ProxyConnectionType = Aws::Crt::Http::AwsHttpProxyConnectionType::Tunneling;
 
     Aws::Crt::ApiHandle apiHandle(allocator);
     Aws::Crt::Mqtt::MqttClient client;
     Aws::Crt::Io::SocketOptions socketOptions;
     std::shared_ptr<Aws::Crt::Mqtt::MqttConnection> connection = client.NewConnection(
-        aws_string_c_str(endpoint), std::stoi(aws_string_c_str(port)), socketOptions, tlsContext, true);
+        aws_string_c_str(endpoint), (uint16_t)std::stoi(aws_string_c_str(port)), socketOptions, tlsContext, true);
     connection->SetHttpProxyOptions(proxyOptions);
     int connectResult = s_ConnectAndDisconnect(connection);
     if (connectResult != AWS_OP_SUCCESS)
